Rewriting syntax highlighting
The tokens produced by haskell-src-exts lexer do not carry enough
context information for proper syntax highlighting.

The solution is to create an abstract presentation tree builder
where tree branches are tagged with classifiers. The problem is
combining a flat list of tokens and comments with the structured
AST.

This will be handled by the TreeBuilder monad which holds a
stack of code fragments ordered by location. We can pop fragments
out of the stack based on the location information stored in an
Annotated AST in order to classify the fragments based on the
AST hierarchy.