Implemented a splitTree function
We need to be able to split the presentation tree into sections
when we run into ProseComment labels.

The splitTree function is a generic function that walks down the
tree until it finds a label that fulfils a test condition. It
breaks the tree in three parts at that point: left, separator and
right. The separator is the tree that starts from the separating
label and left and right are the half-trees on either side.

The spine of the tree all the way to the separation point is
duplicated for both left and right half-trees.