Testing and fixing AST classifiers
Added testClassifer function for debugging the current implementation.

Fixed usage of popPrintableBefore in ASTClassifier instances so that we
mappend instead of (>>). Otherwise the prefix trees are discarded.

Added ASTClassifier instance to Name nodes.

Changed the SourceFragment type-class so that it's possible to return
multiple fragments from an item. This is needed so that we can split
apart some tokens that the haskell-src-exts lexer returns (such as
separating the beginning "{-#" from pragmas).

Added "Pragma" classification.

Changed popPrintablesBefore to also process the leftover span between
the last processed fragment and the parameter span. This fixes some
classified fragments from getting unwanted prefixes.

Changed beforeToTree so that it doesn't create a Leaf from an empty
string.