Workarounds for inaccuracies in location info
haskell-src-exts seems to have some inaccuracies in SrcSpan reporting
for some AST elements. For example for a qualified name (QName) the
ModuleName SrcSpan covers the whole QName, making it impossible to
apply different coloring to the module name part, punctuation and
the actual symbol name.

In order to increase coloring accuracy, popFragments now breaks
fragments into smaller pieces if the position ends up in the middle
of a fragment. There is also a function "popCustom" which lets us
manually tweak fragments when the original SrcSpan info is not
usable.