Fixed formatting problems in qualified names
The current hybrid solution of generating some trees via
type-classes and others via the genericTree function can sometimes
lead to an AST element being processed twice. This is usually not
a problem, because popPrintablesBefore and popPrintables only
operate on fragments within the AST element's SrcSpan.

However, the popCustom hack to fix qualified names always pops a
fragment out of the stack no matter what the current source location
is, so sometimes it ends up mis-classifying fragments that follow
qualified names.

This is fixed by checking our current source location with the new
currentlyWithin function before calling popCustom, but the whole
Classifier module is starting to be in a dire need of a rewrite.